Añadir Clientes v1

+++ api/clientes/delete.php
@@ -1,8 +1,17 @@
 <?php
 header('Content-Type: application/json');
+header('Access-Control-Allow-Origin: *');
+header('Access-Control-Allow-Methods: GET, POST, PUT, DELETE');
+header('Access-Control-Allow-Headers: Content-Type');
 require_once '../db.php';
 
 $data = json_decode(file_get_contents("php://input"), true);
+
+if (!$data || !isset($data['id'])) {
+    echo json_encode(['success' => false, 'error' => 'ID de cliente requerido']);
+    exit;
+}
+
 $id = $data['id'];
 
 $sql = "DELETE FROM clientes_leads WHERE id_cliente = ?";
